Fix uneven display brightness
The display HW is implemented like this:

digit1 --www--+->|--- segA
         R    +->|--- segB
              :  :    :
              +->|--- segP

digit2 --www--+->|--- segA
         R    +->|--- segB
              :  :    :
              +->|--- segP
digit3 --www--+->|--- segA
         R    +->|--- segB
              :  :    :
              +->|--- segP

Notice that there is one current limiting resistor for each display.
This means that the multiplexing have to cycle through the segments
one by one and for each segment apply all the digits in parallel.

1) "segA is active on digit1 and digit3"
2) "segB is active on digit1 and digit2"
3) ...

This this is opposite of the "natural" multiplexing.
Reversing the order of the multiplex fixes the uneven
brightness issue.